Check in of the day's changes so that I can transfer this to my home machine. Very early draft.  Not compiling.

+++ src/util/rational.cpp
@@ -35,7 +35,7 @@ Rational Rational::fromDecimal(const string& dec) {
 
     /* Compute the denominator: 10 raise to the number of decimal places */
     int decPlaces = dec.size() - (i + 1);
-    Integer denominator( Integer(10).pow(decPlaces) );
+    Integer denominator( Integer(10ul).pow(decPlaces) );
 
     return Rational( numerator, denominator );
   } else {
